chore: move common logic to concern
