+++ decidim-bulletin_board-app/app/services/voting_scheme/electionguard/voter.rb
@@ -8,6 +8,8 @@ module VotingScheme
   module Electionguard
     # An implementation of the ElectionGuard voter adapter, using PyCall to execute the ElectionGuard python code
     class Voter < VotingScheme::Voter
+      include Electionguard
+
       def process_message(message_identifier, message)
         results = state.process_message(message_identifier.type, PyCall::Dict.new(message))
 
@@ -24,14 +26,6 @@ module VotingScheme
 
       def state
         @state ||= EGV.Voter.new(SecureRandom.hex)
-      end
-
-      def to_h(dict)
-        return dict unless dict.is_a?(PyCall::Dict)
-
-        dict.inject({}) do |h, (k, v)|
-          h.update(k => to_h(v))
-        end
       end
     end
   end
